requirements.txt
numpy
scipy
pillow
scikit-image
scikit-learn
